web: Review changes refactor AlertOverlay
